requirements/base.txt
# Bleeding edge Django
django==1.8.5

# Configuration
django-environ==0.4.0
django-secure==1.0.1
whitenoise==2.0.4


# Forms
django-braces==1.8.1
django-crispy-forms==1.5.2
django-floppyforms==1.5.2

# Models
django-model-utils==2.3.1

# Images
Pillow==3.0.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.23.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2015.6

# Redis support
django-redis==4.2.0
redis>=2.10.0


# Your custom requirements go here

# DRF
djangorestframework==3.2.4
markdown==2.6.2
django-filter==0.11.0

# mysql
wheel==0.26.0

# this is for linux, if you're using windows, install the wheel
# package from this folder instead
mysql-python==1.2.5

# comments
django-threadedcomments==1.0.1
django-fluent-comments==1.0.5

django-ckeditor==5.0.2

django-rest-swagger==0.3.4

django-grappelli==2.7.2
